+++ README.md
@@ -1,19 +1,18 @@
 # Sprint_5
 автотесты для сервиса Stellar Burgers:
-тест 1 test_registration позволяет проверить успешную регистрацию на рандомных данных(имени,адреса электронной почты и пароля)
-тест 2 test_reg_mistake проверяет, что при невалидном пароле успешная регистрация не происходит
-тест 3 test_transfers проверяет различные переходы по страницам:
+тест 1 test_registration позволяет проверить успешную регистрацию на рандомных данных(имени,адреса электронной почты и пароля), а также проверяет, что при невалидном пароле успешная регистрация не происходит
+тест 2 test_transfers проверяет различные переходы по страницам:
 1) с главной страницы переход по клику на кнопку "Личный кабинет";
 2) из личного кабинета после регистрации по клику на Конструктор;
 3) из личного кабинета после регистрации по клику на логотип;
-тест 4 test_entrance проверяет вход:
+тест 3 test_entrance проверяет вход:
 1) по кнопке «Войти в аккаунт» на главной странице;
 2) вход через кнопку «Личный кабинет»;
 3) вход через кнопку в форме регистрации;
 4) вход через кнопку в форме восстановления пароля;
 тест постоен линейно: вход на главную страницу, клик по интересующей кнопке, заполнение полей логина, пароля,  переход на следующую страницу, клик по интересующей кнопке. Используется только одна фикстура входа на главную страницу.
 тест test_entrance_quit проверяет выход из личного кабинета по кнопке "Выход" после входа 
-тест 5 test_constructor.py прверяет переходы между кнопками Булки, Соусы, Начинки
+тест 4 test_constructor.py прверяет переходы между кнопками Булки, Соусы, Начинки
