--- curl.py
@@ -1,4 +1,0 @@
-main_site = 'https://stellarburgers.nomoreparties.site/'
-login_site = f'{main_site}login'
-profile_site = f'{main_site}account/profile'
-register_site = f'{main_site}register'
