+++ test_transfers.py
@@ -3,7 +3,6 @@ from locators import Locators
 from geniration_em_pas import User
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
-from curl import *
 
 def test_transfers_account(main_page_start):
     main_page_start.find_element(*Locators.entrance_button).click()
